Added Kubernetes YAML files

+++ NitroBoostConsoleService.Core/ProfileService.cs
@@ -21,20 +21,35 @@ public class ProfileService : IProfileService
 
     public async Task DeleteUserInformation(string email)
     {
-        var getResponse = await SendRequest(HttpMethod.Get, $"{_configuration.Audience}users-by-email/{email}");
+        var getResponse = await SendRequest(HttpMethod.Get, $"{_configuration.Audience}users-by-email?email={email}", new Dictionary<string, string>()
+        {
+            {"Accept", "application/json"},
+            {"Authorization", _token}
+        });
         if (!getResponse.IsSuccessStatusCode)
             return;
+
+        var profile = JsonSerializer.Deserialize<List<Profile>>(await getResponse.Content.ReadAsStringAsync());
+        if (profile.Count == 0)
+            return;
         
-        Profile profile = JsonSerializer.Deserialize<Profile>(await getResponse.Content.ReadAsStringAsync());
-        await SendRequest(HttpMethod.Delete, $"{_configuration.Audience}/users/{profile.UserId}");
-        await _service.DeleteUser(email);
+        var deleteResponse = await SendRequest(HttpMethod.Delete, $"{_configuration.Audience}users/{profile[0].UserId}", new Dictionary<string, string>()
+        {
+            {"Authorization", _token}
+        });
+        if (deleteResponse.IsSuccessStatusCode)
+            await _service.DeleteUser(email);
     }
 
-    private async Task<HttpResponseMessage> SendRequest(HttpMethod method, string endpoint, string? body = null)
+    private async Task<HttpResponseMessage> SendRequest(HttpMethod method, string endpoint, Dictionary<string, string>? headers = null, string? body = null)
     {
         var client = new HttpClient();
         var request = new HttpRequestMessage(method, endpoint);
-        request.Headers.Add("Authorization", _token );
+        if (headers == null) return await client.SendAsync(request);
+        foreach (var header in headers)
+        {
+            request.Headers.Add(header.Key, header.Value);
+        }
         return await client.SendAsync(request);
     }
 }
